add register by email
