+++ package.json
@@ -10,13 +10,16 @@
   },
   "dependencies": {
     "@heroui/react": "^2.7.2",
+    "@prisma/adapter-pg": "^6.4.1",
     "@prisma/client": "^6.4.1",
     "@tiptap/pm": "^2.11.5",
     "@tiptap/react": "^2.11.5",
     "@tiptap/starter-kit": "^2.11.5",
+    "bcrypt": "^5.1.1",
     "framer-motion": "^12.4.7",
     "lucide-react": "^0.476.0",
     "next": "15.2.0",
+    "pg": "^8.13.3",
     "prisma": "^6.4.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
