+++ src/app/_actions/is-authenticate.js
@@ -1,0 +1,25 @@
+import { prisma } from "@/libs/postgres";
+import { tryCatch } from "@/utils/try-catch";
+import { cookies } from "next/headers";
+
+export async function isAuthenticateAction() {
+  const { data, error } = await tryCatch(cookies());
+
+  if (error) {
+    console.log(`validateAuthAction [ERROR]: ${error}`);
+  }
+
+  const sessionId = data.get("session_id");
+  if (!sessionId || sessionId.value === "") return { isAuthenticate: false };
+
+  const session = await prisma.session.findUnique({
+    where: { session: sessionId.value },
+  });
+
+  if (!session) return { isAuthenticate: false };
+
+  const now = new Date();
+  if (session.expiresAt < now) return { isAuthenticate: false };
+
+  return { isAuthenticate: true, sessionId: sessionId.value };
+}
